+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - Intensity sum on sum Z-projection for each nucleus as `flat_sumI`.
+- Density profile for each nucleus to `gpseq_anim` and `gpseq_fromfish`.
+- Settings confirmation to `gpseq_fromfish`